requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asn1crypto==1.5.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.0.0
cryptography==44.0.2
filelock==3.18.0
frozenlist==1.5.0
idna==3.10
ijson==3.3.0
multidict==6.2.0
packaging==24.2
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.30.0
sortedcontainers==2.4.0
spider-client==0.1.32
tenacity==9.1.2
tomlkit==0.13.2
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.3.0
yarl==1.19.0
